Camera Animation Part 3 Redux
Once upon a time, there was a camera and that camera got excited for the new fancy animations, what animations (you may ask)?
Well (thank you for asking), because there are BRAND NEW AND FRESH state-of-the-art animations for the enemy when he commits manslaughter (i know it's technically murder, but manslaughter sounds a lot more fancy): one for when cumes from behind and SNAPS the player's neck and one for when he pucks up the player and CHOPS his head.
It even comes with animated reference point for the camera!

Note: I am too lazy to stick em up to the scripts and yada yada yada, do what you need to do