Fantastical new animations!
Prepare for shock and wonder, get ready for the rock and thunder, the bleeding edge, cutting-edge, revolutionary, state-of-the-art, ultra-responsive and compact and in fact very cool as fuck (and every other silicon valley term): new animatioooons! [big audience clapping and cheering], but... [audience gets quiet] ...that's not all [gasp!] , be ready also for a cutscene previeeeeew! [audience dies of a heart attack].